Cambio de números flotantes a números enteros

+++ script/num.js
@@ -10,9 +10,9 @@ function Validar(){
     return Calcular(Inputs);
 }
 function Calcular(Inputs){
-    var N1 = parseFloat(Inputs[0].value);
-    var N2 = parseFloat(Inputs[1].value);
-    var N3 = parseFloat(Inputs[2].value);
+    var N1 = parseInt(Inputs[0].value);
+    var N2 = parseInt(Inputs[1].value);
+    var N3 = parseInt(Inputs[2].value);
     var May;
 
     May = Math.max(N1,N2,N3);
